# static/@typespec/openapi3/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ERC-20 Token API",
    "x-description": "ERC-20 tokens balances and transfers",
    "version": "2.0.0"
  },
  "tags": [
    {
      "name": "Account"
    },
    {
      "name": "Docs"
    },
    {
      "name": "Monitoring"
    }
  ],
  "paths": {
    "/account/balances": {
      "get": {
        "operationId": "Usage_balance",
        "summary": "Account token balances",
        "description": "Token balances of an account.",
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Models.address"
            },
            "explode": false
          },
          {
            "name": "contract",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Models.address"
            },
            "explode": false
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "uint64",
              "default": 10
            },
            "explode": false
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "uint64",
              "default": 1
            },
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "Array of balances.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Balance"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetadata"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Account"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/head": {
      "get": {
        "operationId": "Monitoring_head",
        "summary": "Head block information",
        "description": "Current head block for which data is available (can be lower than head block of the chain).",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Head block information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "block_num": {
                      "type": "integer",
                      "format": "uint32"
                    },
                    "block_hash": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "block_num",
                    "block_hash"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Monitoring"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "Monitoring_health",
        "summary": "Health check",
        "description": "Checks database connection.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK or ApiError.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Monitoring"
        ]
      }
    },
    "/metrics": {
      "get": {
        "operationId": "Monitoring_metrics",
        "summary": "Prometheus metrics",
        "description": "Prometheus metrics.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Metrics as text.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Monitoring"
        ]
      }
    },
    "/openapi": {
      "get": {
        "operationId": "Docs_openapi",
        "summary": "OpenAPI JSON spec",
        "description": "Reflection endpoint to return OpenAPI JSON spec. Also used by Swagger to generate the frontpage.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The OpenAPI JSON spec"
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Docs"
        ]
      }
    },
    "/version": {
      "get": {
        "operationId": "Docs_version",
        "summary": "Api version",
        "description": "Api version and Git short commit hash.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The Api version and commit hash.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Docs"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "status",
          "code",
          "message"
        ],
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              500,
              504,
              400,
              401,
              403,
              404,
              405
            ]
          },
          "code": {
            "type": "string",
            "enum": [
              "bad_database_response",
              "bad_header",
              "missing_required_header",
              "bad_query_input",
              "database_timeout",
              "forbidden",
              "internal_server_error",
              "method_not_allowed",
              "route_not_found",
              "unauthorized",
              "not_found_data"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Balance": {
        "type": "object",
        "required": [
          "last_updated_block",
          "contract",
          "balance"
        ],
        "properties": {
          "last_updated_block": {
            "type": "integer",
            "format": "uint32"
          },
          "contract": {
            "$ref": "#/components/schemas/Models.address"
          },
          "balance": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BalanceChange": {
        "type": "object",
        "required": [
          "block_num",
          "block_hash",
          "timestamp",
          "transaction_id",
          "call_index",
          "log_index",
          "log_block_index",
          "log_ordinal",
          "storage_key",
          "storage_ordinal",
          "from",
          "to",
          "value",
          "contract",
          "owner",
          "old_balance",
          "new_balance",
          "version"
        ],
        "properties": {
          "block_num": {
            "type": "integer",
            "format": "uint32"
          },
          "block_hash": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "transaction_id": {
            "type": "string"
          },
          "call_index": {
            "type": "integer",
            "format": "uint32"
          },
          "log_index": {
            "type": "integer",
            "format": "uint32"
          },
          "log_block_index": {
            "type": "integer",
            "format": "uint32"
          },
          "log_ordinal": {
            "type": "integer",
            "format": "uint64"
          },
          "storage_key": {
            "type": "string"
          },
          "storage_ordinal": {
            "type": "integer",
            "format": "uint64"
          },
          "from": {
            "$ref": "#/components/schemas/Models.address"
          },
          "to": {
            "$ref": "#/components/schemas/Models.address"
          },
          "value": {
            "type": "integer",
            "format": "int64"
          },
          "contract": {
            "$ref": "#/components/schemas/Models.address"
          },
          "owner": {
            "$ref": "#/components/schemas/Models.address"
          },
          "old_balance": {
            "type": "integer",
            "format": "int64"
          },
          "new_balance": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "uint64"
          }
        }
      },
      "BlockRange": {
        "type": "array",
        "items": {
          "type": "integer",
          "format": "uint32"
        },
        "maxItems": 2
      },
      "Models.address": {
        "type": "string",
        "pattern": "^(0[xX])?[0-9a-fA-F]{40}$"
      },
      "Pagination": {
        "type": "object",
        "required": [
          "next_page",
          "previous_page",
          "total_pages",
          "total_results"
        ],
        "properties": {
          "next_page": {
            "type": "integer",
            "format": "int64"
          },
          "previous_page": {
            "type": "integer",
            "format": "int64"
          },
          "total_pages": {
            "type": "integer",
            "format": "int64"
          },
          "total_results": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QueryStatistics": {
        "type": "object",
        "required": [
          "elapsed",
          "rows_read",
          "bytes_read"
        ],
        "properties": {
          "elapsed": {
            "type": "number"
          },
          "rows_read": {
            "type": "integer",
            "format": "int64"
          },
          "bytes_read": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ResponseMetadata": {
        "type": "object",
        "required": [
          "statistics",
          "next_page",
          "previous_page",
          "total_pages",
          "total_results"
        ],
        "properties": {
          "statistics": {
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/QueryStatistics"
              }
            ],
            "nullable": true
          },
          "next_page": {
            "type": "integer",
            "format": "int64"
          },
          "previous_page": {
            "type": "integer",
            "format": "int64"
          },
          "total_pages": {
            "type": "integer",
            "format": "int64"
          },
          "total_results": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Version": {
        "type": "object",
        "required": [
          "version",
          "commit"
        ],
        "properties": {
          "version": {
            "type": "string",
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)$"
          },
          "commit": {
            "type": "string",
            "pattern": "^[0-9a-f]{7}$"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    }
  }
}